requirements.txt
# Core requirements
kagglehub>=0.1.0
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=0.24.0
Pillow>=8.2.0
matplotlib>=3.3.0
seaborn>=0.11.0
tqdm>=4.62.0

# Deep Learning frameworks
torch>=1.9.0
torchvision>=0.10.0

# Object Detection
ultralytics>=8.0.0

# Optional: Detectron2 (requires separate installation)
# detectron2 - See https://detectron2.readthedocs.io/en/latest/tutorials/install.html

# Google Colab specific (if using)
# google-colab